lgammafn_sign was lgamma, moved back to @special
