+++ src/lib/special/gamma/lgammacor.ts
