+++ src/lib/special/gamma/lgammafn_sign.ts
@@ -16,6 +16,8 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
 import { debug } from 'debug';
 
+import { isArray, isEmptyArray, emptyFloat64Array } from '$constants';
+
 import {
     fmod,
     M_LN_SQRT_2PI, // no math alias for this
@@ -25,8 +27,9 @@ import {
 import { ME, ML_ERR_return_NAN, ML_ERROR } from '@common/logger';
 
 import sinpi from '@trig/sinpi';
-import { _gammafn as gammafn } from '../special/gamma/gamma_fn';
+import { _gammafn } from './gamma_fn';
 import { lgammacor } from './lgammacor';
+import type { NumArray } from '$constants';
 
 const { isNaN: ISNAN, POSITIVE_INFINITY: ML_POSINF } = Number;
 const { log, abs: fabs, floor, trunc } = Math;
@@ -35,6 +38,38 @@ const printer_sign = debug('lgammafn_sign');
 
 const xmax = 2.5327372760800758e305;
 const dxrel = 1.490116119384765625e-8;
+
+export function lgammafn<T extends NumArray>(x: T): Float64Array | Float32Array {
+    if (typeof x === 'number') {
+        return new Float64Array([_gammafn(x)]);
+    }
+    if (isEmptyArray(x)) {
+        return emptyFloat64Array;
+    }
+    if (isArray(x)) {
+        if (x instanceof Float64Array) {
+            const rc = new Float64Array(x.length);
+            for (let i = 0; i < x.length; i++) {
+                rc[i] = _gammafn(x[i]);
+            }
+            return rc;
+        }
+        if (x instanceof Float32Array) {
+            const rc = new Float32Array(x.length);
+            for (let i = 0; i < x.length; i++) {
+                rc[i] = _gammafn(x[i]);
+            }
+            return rc;
+        }
+        // array with numbers
+        const rc = new Float64Array(x.length);
+        for (let i = 0; i < x.length; i++) {
+            rc[i] = _gammafn(x[i]);
+        }
+        return rc;
+    }
+    throw new TypeError(`gammafn: argument not of number, number[], Float64Array, Float32Array`);
+}
 
 export function lgammafn_sign(x: number, sgn?: number[]): number {
     //let ans: number;
@@ -75,7 +110,7 @@ export function lgammafn_sign(x: number, sgn?: number[]): number {
     const y = fabs(x);
 
     if (y < 1e-306) return -log(y); // denormalized range, R change
-    if (y <= 10) return log(fabs(gammafn(x) as number));
+    if (y <= 10) return log(fabs(_gammafn(x) as number));
 
     //  ELSE  y = |x| > 10 ----------------------
